requirements.txt
blescan
pygame
paho-mqtt
requests
